removing unused welcome controller, minor tweaks to permission checks - #322

+++ app/Http/Controllers/RoleController.php
@@ -13,7 +13,7 @@ class RoleController extends Controller
     {
         $this->middleware('auth');
     }
-    
+
     public function index()
     {
         $this->authorize('show-role');
@@ -31,7 +31,7 @@ class RoleController extends Controller
         $this->authorize('create-role');
         return view('admin.roles.create');
     }
-    
+
     /**
      * Store a newly created resource in storage.
      *
@@ -51,7 +51,7 @@ class RoleController extends Controller
 
         return Redirect::action('RoleController@index');
     }
-    
+
 
     /**
      * Display the specified resource.
@@ -62,11 +62,11 @@ class RoleController extends Controller
     public function show($id)
     {
         $this->authorize('show-role');
-        
+
         $role = \App\Role::with('users', 'permissions')->findOrFail($id);
         $permissions = \App\Permission::orderBy('name')->pluck('name', 'id');
         $users = \App\User::orderBy('name')->pluck('name', 'id');
-        
+
         return view('admin.roles.show', compact('role', 'permissions', 'users'));//
     }
 
@@ -79,7 +79,7 @@ class RoleController extends Controller
     public function edit($id)
     {
         $this->authorize('update-role');
-        
+
         $role = \App\Role::findOrFail($id);
         return view('admin.roles.edit', compact('role'));//
     }
@@ -94,13 +94,13 @@ class RoleController extends Controller
     public function update(Request $request, $id)
     {
         $this->authorize('update-role');
-        
+
         $role = \App\Role::findOrFail($request->input('id'));
         $role->name= $request->input('name');
         $role->display_name= $request->input('display_name');
         $role->description= $request->input('description');
         $role->save();
-    
+
         return Redirect::action('RoleController@index');
     }
 
@@ -113,25 +113,26 @@ class RoleController extends Controller
     public function destroy($id)
     {
         $this->authorize('delete-role');
-        
+
         \App\Role::destroy($id);
         return Redirect::action('RoleController@index');
     }
 
     public function update_permissions(Request $request)
-    {
+    {   $this->authorize('update-role');
+        $this->authorize('delete-permission');
         $role = \App\Role::findOrFail($request->input('id'));
         $role->permissions()->detach();
         $role->permissions()->sync($request->input('permissions'));
-    
+
         return Redirect::action('RoleController@index');
     }
     public function update_users(Request $request)
-    {
+    {   $this->authorize('update-role');
         $role = \App\Role::findOrFail($request->input('id'));
         $role->users()->detach();
         $role->users()->sync($request->input('users'));
-    
+
         return Redirect::action('RoleController@index');
     }
 }
